requirements.txt
torch>=2.0.0
python-chess>=1.9.0
numpy>=1.21.0
gymnasium>=0.29.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0
PyYAML>=6.0 